ctr: move signals to commands package
