+++ cmd/ctr/commands/signals_linux.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"syscall"
