+++ cmd/ctr/start.go
@@ -70,8 +70,8 @@ var taskStartCommand = cli.Command{
 				logrus.WithError(err).Error("console resize")
 			}
 		} else {
-			sigc := forwardAllSignals(ctx, task)
-			defer stopCatch(sigc)
+			sigc := commands.ForwardAllSignals(ctx, task)
+			defer commands.StopCatch(sigc)
 		}
 
 		status := <-statusC
